requirements.txt
pandas>=2.0
openpyxl>=3.1
rapidfuzz>=3.5
typer>=0.12
pydantic>=2.6
pyyaml>=6.0
rich>=13.7
yt-dlp>=2024.7.0
webvtt-py>=0.5
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0
scikit-learn>=1.3
numpy>=1.24
requests>=2.31
